requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.4
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.3.2
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
firebase-admin==3.2.1
Flask==1.1.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
httplib2==0.17.0
idna==2.8
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==0.6.2
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.1
pandas==1.0.0
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.6.1
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
requests==2.22.0
rsa==4.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==1.9.5
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.15
wcwidth==0.1.9
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.1
yarl==1.4.2
